Bundle type definitions to avoid internal resolution error (#62)
This PR fixes the issue that "[i]mport found in a type declaration file failed
to resolve", as suggested by
https://arethetypeswrong.github.io/?p=tshet-uinh%400.16.0-beta.8.

Because the JS files are bundled, the modules in the unbundled type definition
file don't really corresponds to anything in the bundled JS files. Therefore, we
will also have to bundle the .d.ts file to align it with the bundled JS file.

Another advantage of bundling type definitions is to avoid exposing internal
helpers like ./src/lib/utils.ts, promoting encapsulation.

* Bundle types to avoid internal resolution error

* Make ES module the default & fix missing type declaration file for CJS

* Fix: test files cannot be emitted due to `tsconfig.test.json` errors

* Fix: Test errors due to incorrect module type of test files

* Fix: `typedoc_cf_analytics.mjs` isn't renamed to `.js` in `doc:html` script

* Add missing UMD module declaration

* Remove incorrect exports from type declaration file

* Avoid using multiple-language selectors in `settings.json`

* Remove unused `rollup-plugin-delete` package

* Avoid `rm` command in `package.json`

Found a workaround for emitting type definitions in a subfolder

* Move DTS fix to an individual script

* Place raw type definitions in `build/types` directory

* Remove redundant `tsBuildInfoFile` property in TSConfigs

TypeScript already places the file there.